Draggable Buckets

A custom widget with draggable elements that can be put into buckets.

draggable_buckets(input_id, label, elements, buckets)

Arguments

input_id

(character(1)) the HTML id of this widget

label

(character(1) or shiny.tag) the header of this widget

elements

(character) the elements to drag into buckets

buckets

(character or list) the names of the buckets the elements can be put in or a list of key-pair values where key is a name of a bucket and value is a character vector of elements in a bucket

Value

the HTML code comprising an instance of this widget

Examples

ui <- shiny::fluidPage(
  draggable_buckets("id", "Choices #1", c("a", "b"), c("bucket1", "bucket2")),
  draggable_buckets("id2", "Choices #2", letters, c("vowels", "consonants")),
  shiny::verbatimTextOutput("out")
)
server <- function(input, output) {
  shiny::observeEvent(input$id, {
    print(shiny::isolate(input$id))
  })
  output$out <- shiny::renderPrint(input$id)
}
if (interactive()) shiny::shinyApp(ui, server)

# With default elements in the bucket
ui <- shiny::fluidPage(
  draggable_buckets("id", "Choices #1", c("a", "b"), list(bucket1 = character(), bucket2 = c("c"))),
  shiny::verbatimTextOutput("out")
)
server <- function(input, output) {
  shiny::observeEvent(input$id, {
    print(shiny::isolate(input$id))
  })
  output$out <- shiny::renderPrint(input$id)
}
if (interactive()) shiny::shinyApp(ui, server)