Panel widget

panel_item(title, ..., collapsed = TRUE, input_id = NULL)

Arguments

title

(character)
title of panel

...

content of panel

collapsed

(logical, optional)
whether to initially collapse panel

input_id

(character, optional)
name of the panel item element. If supplied, this will register a shiny input variable that indicates whether the panel item is open or collapsed and is accessed with input$input_id.

Value

(shiny.tag)